+++ release_notes.md
@@ -1,6 +1,7 @@
 ###In Development
   - [#236](https://github.com/Mehdik/Humanizer/pull/236): Added Turkish localisation
   - [#239](https://github.com/Mehdik/Humanizer/pull/239): Added Serbian localisation
+  - [#227](https://github.com/MehdiK/Humanizer/pull/227): Moved localiser registry to their own classes, allowed public access via Configurator, and made the default registrations lazy loaded
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.24.1...master)
 
@@ -19,7 +20,6 @@
   - [#228](https://github.com/Mehdik/Humanizer/pull/228): Fixed the "twenties" in SpanishNumberToWordsConverter
   - [#231](https://github.com/Mehdik/Humanizer/pull/231): Added more settings for FromNow, Dual and Plural (Arabic)
   - [#222](https://github.com/Mehdik/Humanizer/pull/222): Updated Ordinalize and ToOrdinalWords to account for special exceptions with 1 and 3.
-
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.22.1...v1.23.1)